// client/src/app/models/funcionario.model.ts
export interface Funcionario {
  id: number;
  nome: string;
  data_aniversario: Date;
  sexo: string;
  cpf: string;
  numero: string;
  rua: string;
  bairro: string;
  complemento: string;
  cidade: string;
  uf: string;
  cep: string;
  salario: string;
  status: string;
  cargo_id: number;
  filial_id: number;
}